Bug Fixes + Various Changes
Reworded some achievements.

Cleaned up code by adding missing @Override annotations and removing unnecessary casts.

Readded Goblin Drum. Also added achievement and Thaumcraft aspects for the block.

Added sound effect for Scepter of Life.

Fixed Goblin Ranger not spawning from spawners.

Fixed Goblin Spawner death config not working.

Fixed Goblin Miners not spawning from Goblin Miner Spawners.

Decreased the spawn rate of Goblin Bombers from Goblin Spawners because they are more likely to help the player by accidentally killing goblin groups.

Fixed Goblin Rider head shaking when riding Direwolf.

Fixed Goblin Solider moving very slow when the player is far away.

Renamed Goblin Arsonist back to Goblin Bomber.

Fixed Thaumcraft aspects for Goblin Ranger and Goblin Ranger Guard.

Change mod page URL.

+++ src/main/java/goblin/item/GoblinsSpawnEgg.java
@@ -116,7 +116,7 @@ public class GoblinsSpawnEgg extends ItemMonsterPlacer {
 
                     if (world.getBlock(i, j, k) instanceof BlockLiquid)
                     {
-                        Entity entity = spawnCreature(world, itemStack.getItemDamage(), (double)i, (double)j, (double)k);
+                        Entity entity = spawnCreature(world, itemStack.getItemDamage(), i, j, k);
 
                         if (entity != null)
                         {
@@ -162,7 +162,7 @@ public class GoblinsSpawnEgg extends ItemMonsterPlacer {
                 yOffset = 0.5D;
             }
 
-            Entity entity = spawnCreature(world, itemStack.getItemDamage(), (double)x + 0.5D, (double)y + yOffset, (double)z + 0.5D);
+            Entity entity = spawnCreature(world, itemStack.getItemDamage(), x + 0.5D, y + yOffset, z + 0.5D);
 
             if (entity != null)
             {
